refactor(#334): Review 컴포넌트 index 파일 추가 및 import 방식 변경
