+++ src/components/Review/index.ts
@@ -1,0 +1,7 @@
+export { default as Rating } from './Rating';
+
+export { default as Review } from './Review';
+
+export { default as ReviewStatistics } from './ReviewStatistics';
+
+export { default as UserReview } from './UserReview';
